requirements.txt
behave==1.3.3
appium-python-client==5.2.4
behave-html-formatter==0.9.10
python-dotenv==1.1.1




